+++ src/main/java/com/sprint/mission/discodeit/event/kafka/consumer/SseBroadcastKafkaListener.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.event.listener.kafka;
+package com.sprint.mission.discodeit.event.kafka.consumer;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
