+++ src/main/java/com/sprint/mission/discodeit/event/listener/KafkaProduceRequiredEventListener.java
@@ -3,10 +3,9 @@ package com.sprint.mission.discodeit.event.listener;
 import com.sprint.mission.discodeit.event.MessageCreatedEvent;
 import com.sprint.mission.discodeit.event.RoleUpdatedEvent;
 import com.sprint.mission.discodeit.event.S3FileUploadFailedEvent;
-import com.sprint.mission.discodeit.event.publisher.KafkaEventPublisher;
+import com.sprint.mission.discodeit.event.kafka.producer.KafkaEventPublisher;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.event.EventListener;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.event.TransactionPhase;
