+++ src/main/java/com/sprint/mission/discodeit/event/listener/SseNotificationEventListener.java
@@ -1,10 +1,9 @@
 package com.sprint.mission.discodeit.event.listener;
 
 import com.sprint.mission.discodeit.event.SseNotificationEvent;
-import com.sprint.mission.discodeit.event.publisher.SseEventPublisher;
+import com.sprint.mission.discodeit.event.kafka.producer.SseEventPublisher;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.event.EventListener;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.event.TransactionPhase;
